Fix GIF parser
Previously, the GIF parser would assume a GIF is animated if it had the
Netscape Application Extension. This can lead to false positives where a
GIF can have all the indicators of an animated GIF but only one Image
Descriptor.

Now, the code checks for the absence of the Graphic Control Extension
before an Image Descriptor to mean non-animated and otherwise checks if
there are multiple Image Descriptors. This unfortunately requires
parsing much further into the data stream.

The example cli now also supports specifiying the buffer size.